requirements.txt
fitz
langchain_community
openpyxl
langchain_chroma>=0.0.7
langchain_openai>=0.0.9
langchain_huggingface>=0.0.9
openai>=1.0.0
huggingface-hub
langchain>=1.0.0
torch
transformers>=4.35.0
sentence-transformers>=2.2.2
pandas>=2.0.0
pdfplumber>=0.10.0
chromadb>=0.3.33
python-dotenv>=1.0.0
python-docx>=0.8.11
ezdxf>=1.0.0
pywin32>=306
tqdm>=4.65.0
